refactor outcome review supportability router
